updating

+++ resnetconvergence.py
@@ -6,7 +6,7 @@ print("Loading Model...")
 model = models.resnet18().cuda()
 print("Loading Dataset...")
 data=torch.randn((1000,3,224,224)).cuda()
-labels = torch.randint((10),(1000,))
+labels = torch.randint((10),(1000,)).long()
 # imagenet_data = torchvision.datasets.ImageNet(sys.argv[1])
 # data_loader = torch.utils.data.DataLoader(imagenet_data,
 #                                           batch_size=32,
@@ -39,7 +39,7 @@ for epoch in range(500):
 	optimizer.zero_grad()
 	indys=torch.randint(len(data), (32,))
 	batch=data[indys].cuda().half()
-	label = data[indys].cuda().half()
+	label = data[indys].cuda()
 	output = model(batch)
 	if isinstance(output, tuple):
 		output = output[0]
@@ -57,7 +57,7 @@ for epoch in range(500):
 	optimizer.zero_grad()
 	indys=torch.randint(len(data), (32,))
 	batch=data[indys].cuda().bfloat16()
-	label = data[indys].cuda().bfloat16()
+	label = data[indys].cuda()
 	output = model(batch)
 	if isinstance(output, tuple):
 		output = output[0]
